Authorization works

+++ internal/repository/remind_postgres.go
@@ -35,23 +35,24 @@ func (r *RemindPostgres) Create(userID int, rem domain.Remind) (int, error) {
 }
 func (r *RemindPostgres) GetByID(userID int, remindID int) (domain.Remind, error) {
 	var rem domain.Remind
-	queryString := fmt.Sprintf("SELECT t.id, t.title, t.msg, t.remind_date as RemindDate FROM %s t WHERE t.id=$1", remindTable)
-	err := r.db.Get(&rem, queryString, remindID)
+	queryString := fmt.Sprintf("SELECT t.id, t.title, t.msg, t.remind_date as RemindDate FROM %s t WHERE t.id=$1 and t.user_id=$2", remindTable)
+	err := r.db.Get(&rem, queryString, remindID, userID)
 
 	return rem, err
 }
 
 func (r *RemindPostgres) GetAll(userID int) ([]domain.Remind, error) {
 	var rem []domain.Remind
-	queryString := fmt.Sprintf("SELECT t.id, t.title, t.msg, t.remind_date as RemindDate FROM %s t", remindTable)
-	err := r.db.Select(&rem, queryString)
+	queryString := fmt.Sprintf("SELECT t.id, t.title, t.msg, t.remind_date as RemindDate FROM %s t WHERE t.user_id=$1", remindTable)
+	//logrus.Debug("queryString=", queryString, " userID=", userID)
+	err := r.db.Select(&rem, queryString, userID)
 
 	return rem, err
 }
 
 func (r *RemindPostgres) Delete(userID, remindID int) error {
-	queryString := fmt.Sprintf("DELETE FROM %s t WHERE t.id=$1", remindTable)
-	_, err := r.db.Exec(queryString, remindID)
+	queryString := fmt.Sprintf("DELETE FROM %s t WHERE t.id=$1 and t.user_id=$2", remindTable)
+	_, err := r.db.Exec(queryString, remindID, userID)
 	return err
 }
 
@@ -79,8 +80,8 @@ func (r *RemindPostgres) Update(userID, remindID int, input domain.RemindUpdateI
 	}
 
 	updateString := strings.Join(setValues, " ,")
-	queryString := fmt.Sprintf("UPDATE %s t SET %s WHERE id = $%d", remindTable, updateString, argIDs)
-	args = append(args, remindID)
+	queryString := fmt.Sprintf("UPDATE %s t SET %s WHERE id = $%d AND user_id=$%d", remindTable, updateString, argIDs, argIDs+1)
+	args = append(args, remindID, userID)
 
 	logrus.Debugf("updateQuery: %s", queryString)
 	logrus.Debugf("args: %s", args)
